Here we render our first d3 graph using ART!
It's a big commit with a lot going on, so let's dive in.

# weather/WeatherGraph.js
This is the component that actually renders our D3 path using ART.

This is mostly using ART.Surface, ART.Group, and ART.Shape
to render any arbitrary SVG path value in the ART.Shape's d property.

To power that value we have to actually get a proper path value.

# weather/graph-utils.js

This file is what uses d3 to create the SVG path value.

Given a width, height, and data, we can then use that information
to create d3 scale functions and then use those with a d3 shape
generator to create an SVG line path.

This is returned to WeatherGraph and used in the ART.Shape d property.