Here we add animations to our graph!
We can now toggle between showing the temperature Max or Min values.
And when we switch values the graph and numbers animate!

# package.json

To tween the SVG path value we're going to directly use the ART library.
This is only needed for tweening the SVG path.

# weather/ControlButton.js
Generic component to let the user click to change what value to show.

# weather/WeatherPage.js
Hooks up the ControlButton and uses it to change the yAccessor
that is passed to the WeatherGraph.

# weather/WeatherGraph.js
This is where most of the work is done. Most of new changes are commented
inline and will be better understood in the context of where the
changes were applied.